refactor: extract helper to share upvote && collect logic
