+++ lib/groupher_server/cms/cms.ex
@@ -12,9 +12,9 @@ defmodule GroupherServer.CMS do
     ArticleCURD,
     ArticleOperation,
     ArticleEmotion,
-    ArticleReaction,
     ArticleComment,
     ArticleCollect,
+    ArticleUpvote,
     ArticleCommentAction,
     ArticleCommentEmotion,
     CommentCURD,
@@ -82,11 +82,10 @@ defmodule GroupherServer.CMS do
   defdelegate create_content(community, thread, attrs, user), to: ArticleCURD
   defdelegate update_content(content, attrs), to: ArticleCURD
 
-  # ArticleReaction
-  defdelegate upvote_article(thread, article_id, user), to: ArticleReaction
-  defdelegate undo_upvote_article(thread, article_id, user), to: ArticleReaction
+  defdelegate upvote_article(thread, article_id, user), to: ArticleUpvote
+  defdelegate undo_upvote_article(thread, article_id, user), to: ArticleUpvote
 
-  defdelegate upvoted_users(thread, article_id, filter), to: ArticleReaction
+  defdelegate upvoted_users(thread, article_id, filter), to: ArticleUpvote
 
   defdelegate collect_article(thread, article_id, user), to: ArticleCollect
   defdelegate collect_article_ifneed(thread, article_id, user), to: ArticleCollect
